+++ web/src/enterprise.scss
@@ -4,6 +4,7 @@
 
 // Enterprise styles
 @import './enterprise/campaigns/detail/CampaignDetails';
+@import './enterprise/campaigns/global/marketing/CampaignMarketing';
 @import './enterprise/campaigns/list/CampaignNode';
 @import './enterprise/campaigns/detail/changesets/ChangesetNode';
 @import './enterprise/extensions/registry/RegistryArea';
