--- web/src/enterprise/campaigns/global/CampaignsDotComPage.tsx
@@ -1,166 +1,0 @@
-import React from 'react'
-import EyeIcon from 'mdi-react/EyeIcon'
-import GithubCircleIcon from 'mdi-react/GithubCircleIcon'
-import GitlabIcon from 'mdi-react/GitlabIcon'
-import BitbucketIcon from 'mdi-react/BitbucketIcon'
-
-export interface CampaignsDotComPageProps {}
-
-export const CampaignsDotComPage: React.FunctionComponent<CampaignsDotComPageProps> = () => (
-    <>
-        <section className="mt-3 text-center">
-            <h1 className="display-1 font-weight-bold">
-                Large scale
-                <br />
-                code changes
-            </h1>
-            <h1 className="mb-6">Perform and track large-scale campaigns all your repositories.</h1>
-            <img src="/.assets/img/campaigns-splash.png" className="w-100" />
-
-            <a href="https://www.vimeo.com" target="_blank" rel="noopener noreferrer" className="btn btn-primary">
-                <EyeIcon className="icon-inline" /> See it in action
-            </a>
-        </section>
-
-        <section className="pt-3 text-center">
-            <p className="lead">
-                Create and track pull requests across all your repositories, all in one place. Including build and
-                review status, burndown charts, and more. Include existing pull requests and create your patches for
-                repositories, using existing tools.
-            </p>
-        </section>
-
-        <section className="py-5 d-flex">
-            <div className="flex-grow-0 mr-2">
-                <h2>Find and remove vulnerable dependencies.</h2>
-
-                <p>
-                    Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut
-                    labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores
-                    et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet.
-                </p>
-            </div>
-            <div className="flex-grow-0 ml-2">
-                <h2>Roll out company-wide style guides.</h2>
-
-                <p>
-                    Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut
-                    labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores
-                    et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet.
-                    Lorem ipsum dolor sit amet, consetetur sadipscing elitr.
-                </p>
-            </div>
-        </section>
-
-        <section className="py-3 text-center">
-            <h1>Integrates with your favorite code host.</h1>
-            <p className="lead">Campaigns are currently supporting the following code hosts:</p>
-
-            <div className="d-flex justify-content-around mt-4">
-                <div className="flex-grow-0">
-                    <div className="text-center">
-                        <GithubCircleIcon size="4rem" />
-                        <h3>GitHub.com / Enterprise</h3>
-                        <p className="mb-0">Bla bla bla</p>
-                    </div>
-                </div>
-                <div className="flex-grow-0">
-                    <div className="text-center">
-                        <BitbucketIcon size="4rem" />
-                        <h3>Bitbucket Server</h3>
-                        <p className="mb-0">Bla bla bla</p>
-                    </div>
-                </div>
-                <div className="flex-grow-0">
-                    <div className="text-center">
-                        <GitlabIcon size="4rem" />
-                        <h3>Gitlab.com / Exterprise</h3>
-                        <p className="mb-0 text-muted">Coming soon!</p>
-                    </div>
-                </div>
-            </div>
-        </section>
-
-        <section className="py-5 text-center">
-            <h1>What our customers say.</h1>
-            <div className="d-flex mt-5 align-items-stretch">
-                <blockquote className="blockquote flex-basis-100 text-center">
-                    <div className="img-fluid mb-5 mx-auto" style={{ maxWidth: '120px', color: '#ff5a5f' }}>
-                        <svg
-                            viewBox="0 0 2761 991"
-                            version="1.1"
-                            xmlns="http://www.w3.org/2000/svg"
-                            xmlnsXlink="http://www.w3.org/1999/xlink"
-                        >
-                            <g id="airbnb" stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
-                                <g
-                                    transform="translate(130.000000, 105.000000)"
-                                    fill="currentColor"
-                                    fillRule="nonzero"
-                                    id="Shape"
-                                >
-                                    <path d="M1317.62854,196.830214 C1317.62854,224.94903 1294.97121,247.590764 1266.85239,247.590764 C1238.73357,247.590764 1216.09184,224.94903 1216.09184,196.830214 C1216.09184,168.711397 1237.96116,146.069664 1266.85239,146.069664 C1295.76702,146.849875 1317.62854,169.49941 1317.62854,196.830214 Z M1108.31345,299.147129 L1108.31345,311.646115 C1108.31345,311.646115 1084.11129,280.390848 1032.56273,280.390848 C947.441668,280.390848 881.045683,345.218609 881.045683,435.028737 C881.045683,524.066455 946.661457,589.666625 1032.56273,589.666625 C1084.8915,589.666625 1108.31345,557.646752 1108.31345,557.646752 L1108.31345,570.918146 C1108.31345,577.167639 1113.01032,581.841105 1119.25201,581.841105 L1182.50374,581.841105 L1182.50374,288.177358 L1119.25201,288.177358 C1113.01032,288.200764 1108.31345,293.677848 1108.31345,299.147129 Z M1108.31345,487.357512 C1096.61028,504.545567 1073.17273,519.385187 1045.06171,519.385187 C995.081373,519.385187 956.804204,488.145525 956.804204,435.028737 C956.804204,381.919751 995.081373,350.680088 1045.06171,350.680088 C1072.40032,350.680088 1097.39049,366.299919 1108.31345,382.699962 L1108.31345,487.357512 Z M1229.37104,288.200764 L1304.34155,288.200764 L1304.34155,581.864512 L1229.37104,581.864512 L1229.37104,288.200764 Z M2349.32541,280.383046 C2297.77685,280.383046 2273.55909,311.638313 2273.55909,311.638313 L2273.55909,146.849875 L2198.58858,146.849875 L2198.58858,581.864512 L2261.85592,581.864512 C2268.10541,581.864512 2272.77888,576.39523 2272.77888,570.925949 L2272.77888,557.654554 C2272.77888,557.654554 2296.99664,589.674427 2348.52959,589.674427 C2433.66626,589.674427 2500.04664,524.089862 2500.04664,435.052143 C2500.04664,346.014425 2433.66626,280.383046 2349.32541,280.383046 Z M2336.82642,518.597174 C2307.91959,518.597174 2285.27786,503.773158 2273.55909,486.5773 L2273.55909,381.919751 C2285.27786,366.299919 2310.26803,349.899877 2336.82642,349.899877 C2386.80676,349.899877 2425.07613,381.139539 2425.07613,434.248525 C2425.07613,487.357512 2386.81457,518.597174 2336.82642,518.597174 Z M2159.5468,407.705735 L2159.5468,582.644723 L2084.56069,582.644723 L2084.56069,416.28806 C2084.56069,367.875946 2068.94086,348.339454 2026.79384,348.339454 C2004.15211,348.339454 1980.71456,360.058229 1965.85154,377.246285 L1965.85154,581.872314 L1890.89663,581.872314 L1890.89663,288.208566 L1950.2317,288.208566 C1956.4812,288.208566 1961.17027,293.677848 1961.17027,299.147129 L1961.17027,311.646115 C1983.03959,288.988777 2011.93082,280.390848 2040.82204,280.390848 C2073.62993,280.390848 2100.96854,289.776791 2122.83786,308.517467 C2149.38065,330.386791 2159.5468,358.497806 2159.5468,407.705735 Z M1708.89673,280.383046 C1657.36377,280.383046 1633.14601,311.638313 1633.14601,311.638313 L1633.14601,146.849875 L1558.1755,146.849875 L1558.1755,581.864512 L1621.42724,581.864512 C1627.67673,581.864512 1632.3658,576.39523 1632.3658,570.925949 L1632.3658,557.654554 C1632.3658,557.654554 1656.58356,589.674427 1708.11652,589.674427 C1793.25318,589.674427 1859.63356,524.089862 1859.63356,435.052143 C1860.41377,346.006622 1794.03339,280.383046 1708.89673,280.383046 Z M1696.39775,518.597174 C1667.50652,518.597174 1644.86479,503.773158 1633.14601,486.5773 L1633.14601,381.919751 C1644.86479,366.299919 1669.85496,349.899877 1696.39775,349.899877 C1746.39369,349.899877 1784.65525,381.139539 1784.65525,434.248525 C1784.65525,487.357512 1746.39369,518.597174 1696.39775,518.597174 Z M1493.34774,280.383046 C1515.98948,280.383046 1527.70825,284.299707 1527.70825,284.299707 L1527.70825,353.800934 C1527.70825,353.800934 1465.22893,332.711821 1426.18715,377.238483 L1426.18715,582.636921 L1351.20104,582.636921 L1351.20104,288.200764 L1414.46838,288.200764 C1420.71787,288.200764 1425.39134,293.670046 1425.39134,299.139327 L1425.39134,311.638313 C1439.46635,295.230468 1469.9258,280.383046 1493.34774,280.383046 Z M714.704624,555.306117 C710.803568,545.935779 706.894709,535.78523 702.993652,527.187301 C696.736357,513.135694 690.486864,499.848695 685.033187,487.357512 L684.252976,486.5773 C630.355976,369.420765 572.573524,250.719411 511.64682,133.57848 L509.306186,128.881608 C502.898081,116.729582 496.653228,104.492167 490.573312,92.1726639 C482.755594,78.1054534 474.953481,63.2814378 462.454495,49.2142273 C437.464326,17.9745651 401.535593,0.78650921 363.274029,0.78650921 C324.216649,0.78650921 289.07593,17.9745651 263.313352,47.6460025 C251.594577,61.6976088 242.996648,76.5372286 235.194535,90.6044391 C229.099222,102.916209 222.854469,115.153428 216.46166,127.313383 L214.113224,132.010255 C153.974534,249.151186 95.4040683,367.860342 41.5148708,485.009075 L40.7268574,486.561696 C35.265378,499.076286 29.008083,512.347681 22.7585901,526.391485 C18.8575334,534.989414 14.9564767,544.359752 11.05542,554.510302 C0.897068278,583.401528 -2.23157921,610.740133 1.68508173,638.866752 C10.2752086,697.44502 49.3169843,746.629543 103.213984,768.514471 C123.522885,777.1124 144.611998,781.013457 166.47352,781.013457 C172.723012,781.013457 180.525126,780.233246 186.782421,779.445232 C212.560604,776.324387 239.111196,767.742062 264.889379,752.902442 C296.909252,734.934175 327.368703,709.171597 361.72921,671.682441 C396.089718,709.171597 427.32938,734.934175 458.569042,752.902442 C484.355027,767.742062 510.897817,776.324387 536.668198,779.445232 C542.917691,780.241048 550.735408,781.013457 556.984901,781.013457 C578.854225,781.013457 600.715747,777.1124 620.236635,768.514471 C674.913846,746.629543 713.17541,696.664809 721.773339,638.866752 C727.976019,611.535949 724.855174,584.212948 714.704624,555.306117 Z M362.486015,595.916118 C320.30779,542.807132 292.969185,492.826793 283.606649,450.664172 C279.705592,432.695905 278.917578,417.076074 281.266015,403.016665 C282.818635,390.51768 287.515507,379.579117 293.765,370.208778 C308.60462,349.13527 333.594789,335.848271 362.493818,335.848271 C391.400648,335.848271 417.171029,348.347256 431.230437,370.208778 C437.47993,379.579117 442.161198,390.525482 443.737225,403.016665 C446.070057,417.083876 445.289845,433.476116 441.388789,450.664172 C431.987242,492.046582 404.648637,542.034723 362.486015,595.916118 Z M674.086822,632.625062 C668.625343,673.235062 641.286737,708.391385 603.025173,724.011216 C584.284496,731.81333 563.967793,734.161766 543.674496,731.81333 C524.153608,729.464894 504.617116,723.215401 484.323819,711.512231 C456.205002,695.876795 428.093987,671.682441 395.2861,635.745907 C446.826862,572.494173 478.066524,514.688315 489.785298,463.155356 C495.25458,438.937596 496.042593,417.076074 493.686355,396.75937 C490.573312,377.238483 483.535805,359.270215 472.605045,343.650384 C448.379482,308.494061 407.76168,288.200764 362.478213,288.200764 C317.194747,288.200764 276.576944,309.289877 252.366986,343.650384 C241.436226,359.270215 234.398719,377.238483 231.277874,396.75937 C228.157028,417.076074 228.929438,439.717807 235.178931,463.155356 C246.889903,514.688315 278.909776,573.266583 329.670326,636.526118 C297.650453,672.454851 268.751425,696.672611 240.632608,712.292442 C220.323707,724.011216 200.810621,730.260709 181.289733,732.593541 C160.208423,734.941977 139.891719,731.81333 121.931254,724.791428 C83.6696897,709.171597 56.3310842,674.015273 50.8696048,633.405273 C48.5289708,613.884385 50.0815914,594.363497 57.899309,572.494173 C60.232141,564.676456 64.1488019,556.874342 68.0498586,547.504004 C73.5191401,535.005018 79.7530288,521.718019 86.0103238,508.446624 L86.7983372,506.894004 C140.679733,390.51768 198.469987,271.808524 258.608677,156.228015 L260.949311,151.531143 C267.206606,139.827973 273.456099,127.328987 279.69779,115.610213 C285.947283,103.111227 292.969185,91.3924525 301.567114,81.2341009 C317.967156,62.5090286 339.828678,52.3428748 364.046438,52.3428748 C388.264198,52.3428748 410.12572,62.5090286 426.525763,81.2341009 C435.123692,91.4158589 442.145594,103.134633 448.395087,115.610213 C454.652382,127.328987 460.901874,139.827973 467.135763,151.531143 L469.484199,156.228015 C528.705845,272.288756 585.728672,389.458413 640.52213,507.674215 L640.52213,508.454426 C646.779425,520.969016 652.240904,535.01282 658.490397,547.52741 C662.391454,556.882144 666.300313,564.684258 668.640947,572.501975 C674.874835,592.78747 677.215469,612.323962 674.086822,632.625062 Z" />
-                                </g>
-                            </g>
-                        </svg>
-                    </div>
-
-                    <p className="mb-5 mb-md-7">
-                        "Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt
-                        ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo
-                        dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est."
-                    </p>
-
-                    <footer className="blockquote-footer">
-                        <span className="h6 text-uppercase">John Doe</span>
-                    </footer>
-                </blockquote>
-                <div className="border-right mx-2 flex-grow-0 d-block" />
-                <blockquote className="blockquote flex-basis-100 text-center">
-                    <div className="img-fluid mb-5 mx-auto" style={{ maxWidth: '120px', color: '#cb2027' }}>
-                        <svg
-                            viewBox="0 0 2761 991"
-                            version="1.1"
-                            xmlns="http://www.w3.org/2000/svg"
-                            xmlnsXlink="http://www.w3.org/1999/xlink"
-                        >
-                            <g id="pinterest" stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
-                                <g
-                                    transform="translate(131.000000, 177.000000)"
-                                    fill="currentColor"
-                                    fillRule="nonzero"
-                                    id="Shape"
-                                >
-                                    <path d="M574.01881,222.949777 C601.146101,222.949777 627.171838,200.365894 631.699105,172.504233 C637.049513,145.385011 618.662027,122.801128 591.554911,122.801128 C564.439724,122.801128 538.409953,145.385011 533.81409,172.504233 C528.524208,200.365894 546.157149,222.949777 574.01881,222.949777 Z M2414.87365,303.556931 L2475.55598,303.556931 C2483.87211,303.556931 2488.74236,300.179638 2490.61459,292.420337 C2493.51987,280.401339 2496.50718,268.402321 2499.57633,256.42412 C2501.3719,249.439539 2497.30462,244.145621 2489.56146,244.145621 L2429.68209,244.145621 C2433.9955,226.984775 2450.78916,171.269523 2450.78916,134.946472 C2450.78916,62.4052784 2389.19089,0.702089645 2301.7888,0.702089645 C2203.08064,0.702089645 2138.86365,66.9446511 2138.86365,160.157139 C2138.86365,187.344955 2149.00763,219.035829 2161.58069,238.694339 C2156.16991,238.385195 2150.75135,238.231801 2145.33175,238.234349 C2070.32921,238.234349 2023.08745,278.919234 2023.08745,338.120724 C2023.08745,393.682646 2066.88735,415.27392 2108.12503,429.642548 C2144.75474,442.409281 2178.48733,448.348798 2178.48733,476.965004 C2178.48733,496.260364 2159.92634,507.582568 2123.60329,507.582568 C2080.50547,507.582568 2056.82003,487.286528 2050.79982,479.587752 C2054.88324,476.379929 2060.95591,468.741678 2060.95591,456.015294 C2060.95591,438.212883 2048.45145,423.682856 2025.226,423.682856 C2005.57556,423.682856 1988.96347,439.709868 1984.79129,460.792732 C1953.22953,486.612684 1918.78275,503.34582 1869.927,503.34582 C1821.82175,503.34582 1796.85319,478.324798 1796.85319,425.083 C1808.13101,428.516783 1837.49772,432.640551 1856.90606,432.640551 C1938.36056,432.640551 1993.87407,395.63155 1993.87407,321.516707 C1993.87407,281.663032 1958.8866,239.807999 1889.67024,239.807999 L1889.51288,239.807999 C1798.47525,239.807999 1747.48095,297.476189 1725.21583,360.906357 C1718.97784,359.402159 1712.7288,357.94414 1706.46923,356.53242 C1711.44842,342.462382 1715.02343,326.156955 1715.02343,308.080164 C1715.02343,276.75244 1695.62316,239.807999 1643.70887,239.807999 C1607.54722,239.807999 1573.2457,263.493437 1549.4715,298.993349 C1555.3868,275.836495 1559.30882,260.414733 1559.56303,259.381774 C1561.11247,253.325242 1558.59059,246.578725 1549.14466,246.578725 L1495.86251,246.578725 C1488.26865,246.578725 1482.91017,249.261999 1480.75145,257.908999 C1479.30288,263.675012 1460.9961,335.506045 1443.17351,405.4487 C1431.94815,449.058958 1373.87645,503.34582 1300.88738,503.34582 C1252.77406,503.34582 1227.82567,476.678519 1227.82567,423.416546 C1240.41486,428.428013 1268.46617,432.636516 1287.86644,432.636516 C1372.63368,432.636516 1424.83444,395.627515 1424.83444,321.512672 C1424.83444,281.658997 1389.85101,239.803964 1320.62659,239.803964 L1320.47729,239.803964 C1217.96413,239.803964 1155.16746,319.426578 1144.52717,409.705623 C1140.10078,447.235139 1108.42604,498.770133 1074.08821,498.770133 C1057.35911,498.770133 1048.03019,488.262998 1048.03019,470.424272 C1048.03019,455.01865 1058.87627,417.347909 1070.76337,370.138433 C1074.74592,354.345451 1080.79034,330.252478 1087.49247,303.556931 L1143.22387,303.556931 C1151.51982,303.556931 1156.38603,300.179638 1158.27037,292.420337 C1160.93751,281.247428 1165.83196,261.754353 1167.22,256.42412 C1169.01557,249.439539 1164.97251,244.145621 1157.20513,244.145621 L1102.40179,244.145621 C1102.40179,244.145621 1127.63263,143.472422 1128.23385,140.885989 C1130.66292,130.649199 1122.09662,124.741962 1113.13489,126.626306 C1113.13489,126.626306 1070.77548,134.918227 1062.91934,136.556436 C1055.02285,138.154295 1048.89368,142.491918 1046.04094,153.874647 L1023.60635,244.145621 L979.600664,244.145621 C971.304709,244.145621 966.426396,247.530985 964.566262,255.274146 C961.642725,267.290329 958.659479,279.291906 955.616636,291.278433 C953.796852,298.263014 957.860095,303.556931 965.619396,303.556931 L1008.7737,303.556931 C1008.46704,304.791641 993.186502,362.919821 980.827304,414.97533 C975.008837,440.008457 955.523832,498.584523 923.615068,498.584523 C904.553737,498.584523 896.499881,488.969123 896.499881,468.346248 C896.499881,435.868549 928.46917,354.567376 928.46917,317.675389 C928.46917,268.395959 902.031864,239.844314 847.870086,239.844314 C813.729968,239.844314 778.330931,261.887507 763.328809,281.235323 C763.328809,281.235323 767.880286,265.397956 769.437796,259.309144 C771.100215,252.877357 767.650292,246.55855 759.205041,246.55855 L706.249728,246.55855 C694.979978,246.55855 691.86496,252.582802 690.210611,259.208269 C689.548871,261.867332 670.467366,336.797245 652.148475,408.729153 C639.619806,457.984373 609.175747,499.141352 576.625418,499.141352 C559.888247,499.141352 552.435605,488.638253 552.435605,470.799527 C552.435605,455.385835 562.39398,417.351944 574.277049,370.146503 C588.734447,312.676027 601.489076,265.349536 602.852905,259.821589 C604.632339,252.736132 601.464866,246.562585 592.333665,246.562585 L539.059587,246.562585 C529.407872,246.562585 525.961984,251.626508 523.928345,258.663544 C523.928345,258.663544 508.934292,315.42386 493.056576,378.611928 C481.524551,424.46968 468.790098,471.178817 468.790098,493.12517 C468.790098,532.288861 486.374619,561.808906 534.717929,561.808906 C572.029556,561.808906 601.557671,542.8888 624.105239,518.779687 C620.792506,531.828871 618.678167,540.088512 618.464312,540.931826 C616.519443,548.392538 618.875882,554.985724 627.664107,554.985724 L682.180965,554.985724 C691.67128,554.985724 695.318918,551.168616 697.300102,542.892835 C699.208656,534.996344 739.925821,374.84324 739.925821,374.84324 C750.68716,331.822092 777.306042,303.335007 814.754858,303.335007 C832.512884,303.335007 847.882191,315.092991 846.102757,337.943184 C844.125608,363.081221 813.810668,453.388511 813.810668,493.024296 C813.810668,523.024505 824.935158,561.837151 881.408989,561.837151 C919.90287,561.837151 948.24066,543.38107 968.847395,519.562477 C976.392841,544.841739 996.382221,561.837151 1033.04825,561.837151 C1093.98479,561.837151 1130.71537,525.695674 1152.29454,489.392797 C1170.20186,531.570631 1212.14163,561.816976 1273.64307,561.816976 C1336.32676,561.816976 1384.83551,535.391774 1419.21369,499.468187 L1408.4927,541.520936 C1406.5438,549.094627 1409.84039,555.050284 1418.73353,555.050284 L1472.68952,555.050284 C1480.0776,555.050284 1485.3352,551.293701 1487.25183,543.5344 C1488.20005,539.725362 1495.99567,509.596032 1507.00718,466.784704 C1528.06583,384.86214 1561.91139,299.868943 1612.76851,299.868943 C1630.70407,299.868943 1638.05988,309.895913 1638.05988,325.69293 C1638.05988,333.141536 1635.89712,339.577358 1634.02891,343.027281 C1609.37911,338.128794 1589.5471,350.298348 1589.5471,376.618639 C1589.5471,393.767381 1607.59161,409.354578 1631.94685,409.354578 C1650.55626,409.354578 1665.94171,404.790995 1678.21214,396.531354 C1690.28966,399.022486 1702.3402,401.642695 1714.36169,404.39153 C1712.51262,416.270627 1711.56706,428.272844 1711.53315,440.294942 C1711.53315,506.049269 1758.48439,561.812941 1842.64234,561.812941 C1908.59035,561.812941 1953.84689,537.687688 1992.91374,506.759429 C2011.69262,537.485938 2059.10788,562.317315 2121.31544,562.317315 C2206.57495,562.317315 2254.93036,517.984793 2254.93036,459.384518 C2254.93036,406.344469 2211.3887,386.823149 2166.48724,370.570177 C2129.98665,357.383804 2099.54259,350.213613 2099.54259,324.417871 C2099.54259,303.274482 2116.18696,293.602591 2144.71439,293.602591 C2163.22293,293.602591 2177.11946,297.589169 2184.53982,299.671228 C2191.89966,318.090994 2207.12774,338.487909 2233.12524,338.487909 C2255.12404,338.487909 2265.00172,321.088998 2265.00172,305.066021 C2265.00172,259.454404 2189.02271,254.672931 2189.02271,156.852476 C2189.02271,96.0248814 2223.44124,45.8456468 2294.72351,45.8456468 C2344.25311,45.8456468 2376.59766,77.2540709 2376.59766,127.94575 C2376.59766,169.740259 2351.2498,244.141586 2351.2498,244.141586 L2308.16409,244.141586 C2299.88428,244.141586 2294.99789,247.52695 2293.12565,255.270111 C2290.20955,267.289472 2287.2209,279.29112 2284.15989,291.274398 C2282.34818,298.258979 2286.41142,303.552896 2294.17475,303.552896 L2336.24364,303.552896 C2336.24364,303.552896 2292.75443,463.944065 2292.75443,493.113065 C2292.75443,532.276756 2314.82587,561.788731 2363.18936,561.788731 C2431.59468,561.788731 2474.34145,517.335158 2492.92665,452.694491 C2494.37118,447.695128 2491.53861,443.623816 2486.47469,443.623816 L2460.37632,443.623816 C2455.16714,443.623816 2452.14492,446.593574 2450.82951,451.685741 C2444.93438,474.632775 2430.32366,498.774168 2401.55412,498.774168 C2384.82502,498.774168 2375.50014,488.267033 2375.50014,470.432342 C2375.50014,455.022685 2385.70061,418.945768 2398.24946,370.146503 C2404.22932,346.908949 2414.87365,303.556931 2414.87365,303.556931 Z M1875.49933,300.817168 C1899.68914,300.817168 1911.18082,316.90874 1911.18082,336.087085 C1911.18082,368.524434 1886.02261,388.303994 1846.67331,388.303994 C1832.59116,388.303994 1812.90441,384.96705 1801.27958,381.093452 C1806.16596,351.262712 1827.75724,300.817168 1875.49933,300.817168 Z M1306.46374,300.817168 C1332.9414,300.817168 1342.13313,316.90874 1342.13313,336.087085 C1342.13313,368.524434 1316.98702,388.303994 1277.64175,388.303994 C1263.55558,388.303994 1243.86882,384.96705 1232.24803,381.093452 C1237.10616,351.262712 1256.25223,300.817168 1306.46374,300.817168 Z M259.640014,0 C87.030871,0 0,123.753387 0,226.952495 C0,289.438474 23.657193,345.02864 74.3972924,365.740285 C82.7174581,369.145823 90.1700994,365.8573 92.5830282,356.6454 C94.2615873,350.270103 98.2320253,334.190636 100.003389,327.488504 C102.440528,318.377479 101.492304,315.181761 94.778067,307.240885 C80.1471644,289.983198 70.7980742,267.641415 70.7980742,235.994926 C70.7980742,144.182582 139.489879,61.9896736 249.669534,61.9896736 C347.235784,61.9896736 400.832662,121.602733 400.832662,201.217278 C400.832662,305.96986 354.478606,394.3807 285.653645,394.3807 C247.648,394.3807 219.197229,362.948066 228.316324,324.397696 C239.235029,278.374509 260.386488,228.703684 260.386488,195.483546 C260.386488,165.745611 244.424036,140.942479 211.389508,140.942479 C172.536513,140.942479 141.325803,181.135093 141.325803,234.978106 C141.325803,269.271554 152.914318,292.464722 152.914318,292.464722 C152.914318,292.464722 113.153448,460.929922 106.185006,490.433827 C92.3046133,549.191467 104.094877,621.220216 105.091522,628.495317 C105.680632,632.804695 111.220684,633.829584 113.730452,630.577376 C117.313531,625.900813 163.582852,568.773312 179.319344,511.686161 C183.769947,495.521959 204.872986,411.828032 204.872986,411.828032 C217.49446,435.904864 254.386446,457.112814 293.618731,457.112814 C410.403677,457.112814 489.642967,350.641323 489.642967,208.125195 C489.642967,100.362504 398.367278,0 259.640014,0 Z" />
-                                </g>
-                            </g>
-                        </svg>
-                    </div>
-
-                    <p className="mb-5">
-                        "Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt
-                        ut labore et dolore magna aliquyam erat. Stet clita kasd gubergren, no sea takimata sanctus
-                        est."
-                    </p>
-
-                    <footer className="blockquote-footer mb-0">
-                        <span className="h6 text-uppercase">Max Power</span>
-                    </footer>
-                </blockquote>
-            </div>
-        </section>
-
-        <section className="my-3 text-center">
-            <h1>Start using campaigns today.</h1>
-
-            <p className="lead">
-                To use campaigns, a private Sourcegraph instance is required. If your organization has one set up
-                already, just navigate to Campaigns from the top navigation.
-            </p>
-
-            <a href="https://about.sourcegraph.com" className="btn btn-primary">
-                Get started with my instance
-            </a>
-        </section>
-    </>
-)
