+++ web/src/enterprise/campaigns/global/marketing/CampaignMarketing.scss
@@ -1,0 +1,5 @@
+.campaign-marketing {
+    &--rate-btn {
+        font-size: 4rem;
+    }
+}
